emergency/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==3.2.10
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-environ==0.9.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
mailjet-rest==1.3.4
Pillow==9.0.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
